Kumis_Server/requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
Pillow>=10.0.0
tqdm>=4.65.0
requests>=2.31.0  # For Unsplash download

# Code quality (optional)
flake8>=6.0.0
